feat(extend): unlimited seed extension via -l -1
Add an unlimited seed-extension mode: `-l -1` extends the DP across the
whole query rather than a fixed number of nucleotides, capped at MAX_EXT
(256) per side by the fixed DP buffers.

- CLI: `-l/--extension` is now i32 with range -1..=256 and allow_hyphen_values.
- config: ExtendConfig.max_extension is i32; negative means unlimited.
- dp: DpConfig gains an `unlimited` flag and `side_cap()`, which sizes each
  extension window to the query bases available on that side (capped at
  MAX_EXT). The grid is pre-sized to the ceiling so unlimited runs skip regrow.
- search: per-side windows (ext_left/ext_right) replace the single fixed cap.

A query longer than MAX_EXT is rejected up front rather than silently
clamped: `-l -1` promises to span the whole query, and past the cap it
cannot. This mirrors clap already rejecting an explicit `-l > 256`, so you
can neither ask for more than the cap nor be silently given less.

Note for reviewers: `-l -1` is deliberately NOT equivalent to a large fixed
window such as `-l 256`. Unlimited bounds each window by query length, while
a fixed window permits large target-side bulges up to its size, so the two
diverge on real data. Extension is also effectively a no-op at the default
per-nucleotide penalty (3.5); it only does work at a low penalty.

Tests: tests/extension_unlimited.rs covers reach (extension crossing a
mismatch the exact-match seeder cannot, to the query 3' end), rejection of
queries longer than the ceiling, and determinism. The metamorphic approach
is used because risearch2 (C) has no unlimited mode to serve as an oracle.

+++ src/config/mod.rs
@@ -164,8 +164,10 @@ pub struct ScoreConfig {
 /// Arguments for extension strategy.
 #[derive(Debug, Clone)]
 pub struct ExtendConfig {
-    /// Max extension length on the seed (do DP for max this length up- and downstream of seed)
-    pub max_extension: u8,
+    /// Max extension length on the seed (do DP for max this length up- and downstream of seed).
+    /// Negative = unlimited: extend across the whole query. Queries longer than
+    /// MAX_EXT per side are rejected rather than silently clamped.
+    pub max_extension: i32,
 }
 
 /// Hit acceptance and pruning policies.
